+++ apps/sim/app/workspace/[workspaceId]/w/components/search-modal/search-modal.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 import * as DialogPrimitive from '@radix-ui/react-dialog'
 import * as VisuallyHidden from '@radix-ui/react-visually-hidden'
 import { BookOpen, Building2, LibraryBig, ScrollText, Search, Shapes, Workflow } from 'lucide-react'
@@ -13,8 +13,9 @@ import {
 } from '@/app/workspace/[workspaceId]/templates/components/template-card'
 import { getKeyboardShortcutText } from '@/app/workspace/[workspaceId]/w/hooks/use-keyboard-shortcuts'
 import { getAllBlocks } from '@/blocks'
+import { useSearchNavigation } from './hooks/use-search-navigation'
 
-interface SearchModalProps {
+export interface SearchModalProps {
   open: boolean
   onOpenChange: (open: boolean) => void
   templates?: TemplateData[]
@@ -24,7 +25,7 @@ interface SearchModalProps {
   isOnWorkflowPage?: boolean
 }
 
-interface TemplateData {
+export interface TemplateData {
   id: string
   title: string
   description: string
@@ -39,21 +40,21 @@ interface TemplateData {
   isStarred?: boolean
 }
 
-interface WorkflowItem {
+export interface WorkflowItem {
   id: string
   name: string
   href: string
   isCurrent?: boolean
 }
 
-interface WorkspaceItem {
+export interface WorkspaceItem {
   id: string
   name: string
   href: string
   isCurrent?: boolean
 }
 
-interface BlockItem {
+export interface BlockItem {
   id: string
   name: string
   icon: React.ComponentType<any>
@@ -61,7 +62,7 @@ interface BlockItem {
   type: string
 }
 
-interface ToolItem {
+export interface ToolItem {
   id: string
   name: string
   icon: React.ComponentType<any>
@@ -69,7 +70,7 @@ interface ToolItem {
   type: string
 }
 
-interface PageItem {
+export interface PageItem {
   id: string
   name: string
   icon: React.ComponentType<any>
@@ -77,7 +78,7 @@ interface PageItem {
   shortcut?: string
 }
 
-interface DocItem {
+export interface DocItem {
   id: string
   name: string
   icon: React.ComponentType<any>
@@ -85,256 +86,17 @@ interface DocItem {
   type: 'main' | 'block' | 'tool'
 }
 
-// Navigation types for the new system
-interface NavigationPosition {
+export interface NavigationPosition {
   sectionIndex: number
   itemIndex: number
 }
 
-interface NavigationSection {
+export interface NavigationSection {
   id: string
   name: string
   type: 'grid' | 'list'
   items: any[]
   gridCols?: number // How many columns per row for grid sections
-}
-
-// Custom hook for navigation logic
-function useSearchNavigation(sections: NavigationSection[], open: boolean) {
-  const [position, setPosition] = useState<NavigationPosition>({ sectionIndex: 0, itemIndex: 0 })
-  const scrollRefs = useRef<Map<string, HTMLElement>>(new Map())
-  const lastItemIndex = useRef<Map<string, number>>(new Map()) // Store last item index for each section
-
-  // Reset position when modal opens or sections change
-  useEffect(() => {
-    if (open) {
-      setPosition({ sectionIndex: 0, itemIndex: 0 })
-    }
-  }, [open, sections])
-
-  // Get current selected item
-  const getCurrentItem = useCallback(() => {
-    const section = sections[position.sectionIndex]
-    if (!section || position.itemIndex >= section.items.length) return null
-
-    return {
-      section,
-      item: section.items[position.itemIndex],
-      position,
-    }
-  }, [sections, position])
-
-  // Navigate with keyboard
-  const navigate = useCallback(
-    (direction: 'up' | 'down' | 'left' | 'right') => {
-      setPosition((prev) => {
-        const section = sections[prev.sectionIndex]
-        if (!section) return prev
-
-        // Debug logging
-        console.log(
-          'Navigate:',
-          direction,
-          'Current:',
-          prev,
-          'Section:',
-          section.id,
-          'Items:',
-          section.items.length
-        )
-
-        switch (direction) {
-          case 'down':
-            if (section.type === 'grid' && section.gridCols) {
-              // CSS Grid with grid-flow-col and 2 rows means items flow column-wise
-              // Index pattern: 0=R0C0, 1=R1C0, 2=R0C1, 3=R1C1, 4=R0C2, 5=R1C2...
-              const totalRows = section.id === 'templates' ? 1 : 2
-              const currentCol = Math.floor(prev.itemIndex / totalRows)
-              const currentRow = prev.itemIndex % totalRows
-
-              console.log('Grid down:', {
-                itemIndex: prev.itemIndex,
-                totalRows,
-                currentCol,
-                currentRow,
-              })
-
-              // Try to move down one row in the same column
-              if (currentRow < totalRows - 1) {
-                const nextIndex = currentCol * totalRows + (currentRow + 1)
-                if (nextIndex < section.items.length) {
-                  return { ...prev, itemIndex: nextIndex }
-                }
-              }
-            } else if (section.type === 'list') {
-              // For list navigation, move to next item within the list
-              if (prev.itemIndex < section.items.length - 1) {
-                return { ...prev, itemIndex: prev.itemIndex + 1 }
-              }
-            }
-            // Move to next section
-            if (prev.sectionIndex < sections.length - 1) {
-              const nextSection = sections[prev.sectionIndex + 1]
-
-              // Store current item index for this section
-              lastItemIndex.current.set(section.id, prev.itemIndex)
-
-              // Check if we have a remembered position for the next section
-              const rememberedIndex = lastItemIndex.current.get(nextSection.id)
-              const targetIndex =
-                rememberedIndex !== undefined
-                  ? Math.min(rememberedIndex, nextSection.items.length - 1)
-                  : 0
-
-              console.log('Section transition down:', {
-                from: section.id,
-                to: nextSection.id,
-                storedIndex: prev.itemIndex,
-                rememberedIndex,
-                targetIndex,
-              })
-
-              return { sectionIndex: prev.sectionIndex + 1, itemIndex: targetIndex }
-            }
-            return prev
-
-          case 'up':
-            if (section.type === 'grid' && section.gridCols) {
-              // CSS Grid with grid-flow-col and 2 rows means items flow column-wise
-              const totalRows = section.id === 'templates' ? 1 : 2
-              const currentCol = Math.floor(prev.itemIndex / totalRows)
-              const currentRow = prev.itemIndex % totalRows
-
-              console.log('Grid up:', {
-                itemIndex: prev.itemIndex,
-                totalRows,
-                currentCol,
-                currentRow,
-              })
-
-              // Try to move up one row in the same column
-              if (currentRow > 0) {
-                const prevIndex = currentCol * totalRows + (currentRow - 1)
-                return { ...prev, itemIndex: prevIndex }
-              }
-            } else if (section.type === 'list') {
-              // For list navigation, move to previous item within the list
-              if (prev.itemIndex > 0) {
-                return { ...prev, itemIndex: prev.itemIndex - 1 }
-              }
-            }
-            // Move to previous section
-            if (prev.sectionIndex > 0) {
-              const prevSection = sections[prev.sectionIndex - 1]
-
-              // Store current item index for this section
-              lastItemIndex.current.set(section.id, prev.itemIndex)
-
-              // Check if we have a remembered position for the previous section
-              const rememberedIndex = lastItemIndex.current.get(prevSection.id)
-              const targetIndex =
-                rememberedIndex !== undefined
-                  ? Math.min(rememberedIndex, prevSection.items.length - 1)
-                  : prevSection.items.length - 1 // Default to last item
-
-              console.log('Section transition up:', {
-                from: section.id,
-                to: prevSection.id,
-                storedIndex: prev.itemIndex,
-                rememberedIndex,
-                targetIndex,
-              })
-
-              return { sectionIndex: prev.sectionIndex - 1, itemIndex: targetIndex }
-            }
-            return prev
-
-          case 'right':
-            if (section.type === 'grid' && section.gridCols) {
-              // CSS Grid with grid-flow-col: move right means next column, same row
-              const totalRows = section.id === 'templates' ? 1 : 2
-              const currentCol = Math.floor(prev.itemIndex / totalRows)
-              const currentRow = prev.itemIndex % totalRows
-              const totalCols = Math.ceil(section.items.length / totalRows)
-
-              // Can we move right to the next column?
-              if (currentCol < totalCols - 1) {
-                const nextIndex = (currentCol + 1) * totalRows + currentRow
-                if (nextIndex < section.items.length) {
-                  return { ...prev, itemIndex: nextIndex }
-                }
-              }
-            } else if (section.type === 'list') {
-              // List navigation - just move to next item
-              if (prev.itemIndex < section.items.length - 1) {
-                return { ...prev, itemIndex: prev.itemIndex + 1 }
-              }
-            }
-            return prev
-
-          case 'left':
-            if (section.type === 'grid' && section.gridCols) {
-              // CSS Grid with grid-flow-col: move left means previous column, same row
-              const totalRows = section.id === 'templates' ? 1 : 2
-              const currentCol = Math.floor(prev.itemIndex / totalRows)
-              const currentRow = prev.itemIndex % totalRows
-
-              // Can we move left to the previous column?
-              if (currentCol > 0) {
-                const prevIndex = (currentCol - 1) * totalRows + currentRow
-                return { ...prev, itemIndex: prevIndex }
-              }
-            } else if (section.type === 'list') {
-              // List navigation - move to previous item
-              if (prev.itemIndex > 0) {
-                return { ...prev, itemIndex: prev.itemIndex - 1 }
-              }
-            }
-            return prev
-
-          default:
-            return prev
-        }
-      })
-    },
-    [sections]
-  )
-
-  // Scroll item into view
-  const scrollIntoView = useCallback(() => {
-    const current = getCurrentItem()
-    if (!current) return
-
-    const container = scrollRefs.current.get(current.section.id)
-    if (!container) return
-
-    const itemSelector = `[data-nav-item="${current.section.id}-${current.position.itemIndex}"]`
-    const element = container.querySelector(itemSelector) as HTMLElement
-    if (!element) return
-
-    // Use modern scrollIntoView with proper options
-    element.scrollIntoView({
-      behavior: 'smooth',
-      block: 'nearest',
-      inline: 'center', // This ensures horizontal centering
-    })
-  }, [getCurrentItem])
-
-  // Trigger scroll when position changes
-  useEffect(() => {
-    if (open) {
-      // Small delay to ensure DOM is updated
-      const timer = setTimeout(scrollIntoView, 10)
-      return () => clearTimeout(timer)
-    }
-  }, [position, open, scrollIntoView])
-
-  return {
-    position,
-    navigate,
-    getCurrentItem,
-    scrollRefs,
-  }
 }
 
 export function SearchModal({
@@ -351,15 +113,12 @@ export function SearchModal({
   const router = useRouter()
   const workspaceId = params.workspaceId as string
 
-  // Local state for templates to handle star changes
   const [localTemplates, setLocalTemplates] = useState<TemplateData[]>(templates)
 
-  // Update local templates when props change
   useEffect(() => {
     setLocalTemplates(templates)
   }, [templates])
 
-  // Get all available blocks - only when on workflow page
   const blocks = useMemo(() => {
     if (!isOnWorkflowPage) return []
 
@@ -383,7 +142,6 @@ export function SearchModal({
       .sort((a, b) => a.name.localeCompare(b.name))
   }, [isOnWorkflowPage])
 
-  // Get all available tools - only when on workflow page
   const tools = useMemo(() => {
     if (!isOnWorkflowPage) return []
 
@@ -402,7 +160,6 @@ export function SearchModal({
       .sort((a, b) => a.name.localeCompare(b.name))
   }, [isOnWorkflowPage])
 
-  // Define pages
   const pages = useMemo(
     (): PageItem[] => [
       {
@@ -435,12 +192,10 @@ export function SearchModal({
     [workspaceId]
   )
 
-  // Define docs
   const docs = useMemo((): DocItem[] => {
     const allBlocks = getAllBlocks()
     const docsItems: DocItem[] = []
 
-    // Add individual block/tool docs
     allBlocks.forEach((block) => {
       if (block.docsLink) {
         docsItems.push({
@@ -456,7 +211,6 @@ export function SearchModal({
     return docsItems.sort((a, b) => a.name.localeCompare(b.name))
   }, [])
 
-  // Filter all items based on search query
   const filteredBlocks = useMemo(() => {
     if (!searchQuery.trim()) return blocks
     const query = searchQuery.toLowerCase()
@@ -505,11 +259,9 @@ export function SearchModal({
     return docs.filter((doc) => doc.name.toLowerCase().includes(query))
   }, [docs, searchQuery])
 
-  // Create navigation sections
   const navigationSections = useMemo((): NavigationSection[] => {
     const sections: NavigationSection[] = []
 
-    // Add blocks section
     if (filteredBlocks.length > 0) {
       sections.push({
         id: 'blocks',
@@ -520,7 +272,6 @@ export function SearchModal({
       })
     }
 
-    // Add tools section
     if (filteredTools.length > 0) {
       sections.push({
         id: 'tools',
@@ -531,7 +282,6 @@ export function SearchModal({
       })
     }
 
-    // Add templates section
     if (filteredTemplates.length > 0) {
       sections.push({
         id: 'templates',
@@ -542,7 +292,6 @@ export function SearchModal({
       })
     }
 
-    // Add list sections
     const listItems = [
       ...filteredWorkspaces.map((item) => ({ type: 'workspace', data: item })),
       ...filteredWorkflows.map((item) => ({ type: 'workflow', data: item })),
@@ -570,17 +319,14 @@ export function SearchModal({
     filteredDocs,
   ])
 
-  // Use the navigation hook
   const { navigate, getCurrentItem, scrollRefs } = useSearchNavigation(navigationSections, open)
 
-  // Clear search when modal closes
   useEffect(() => {
     if (!open) {
       setSearchQuery('')
     }
   }, [open])
 
-  // Handle block/tool click
   const handleBlockClick = useCallback(
     (blockType: string) => {
       const event = new CustomEvent('add-block-from-toolbar', {
@@ -592,7 +338,6 @@ export function SearchModal({
     [onOpenChange]
   )
 
-  // Handle page navigation
   const handlePageClick = useCallback(
     (href: string) => {
       if (href.startsWith('http')) {
@@ -605,7 +350,6 @@ export function SearchModal({
     [router, onOpenChange]
   )
 
-  // Handle workflow/workspace navigation
   const handleNavigationClick = useCallback(
     (href: string) => {
       router.push(href)
@@ -614,7 +358,6 @@ export function SearchModal({
     [router, onOpenChange]
   )
 
-  // Handle docs navigation
   const handleDocsClick = useCallback(
     (href: string) => {
       if (href.startsWith('http')) {
@@ -627,7 +370,6 @@ export function SearchModal({
     [router, onOpenChange]
   )
 
-  // Handle item selection
   const handleItemSelection = useCallback(() => {
     const current = getCurrentItem()
     if (!current) return
@@ -671,7 +413,6 @@ export function SearchModal({
     onOpenChange,
   ])
 
-  // Handle keyboard navigation
   useEffect(() => {
     if (!open) return
 
@@ -707,7 +448,6 @@ export function SearchModal({
     return () => document.removeEventListener('keydown', handleKeyDown)
   }, [open, navigate, handleItemSelection, onOpenChange])
 
-  // Handle template star changes
   const handleStarChange = useCallback(
     (templateId: string, isStarred: boolean, newStarCount: number) => {
       setLocalTemplates((prevTemplates) =>
@@ -719,7 +459,6 @@ export function SearchModal({
     []
   )
 
-  // Helper to check if item is selected
   const isItemSelected = useCallback(
     (sectionId: string, itemIndex: number) => {
       const current = getCurrentItem()
@@ -728,7 +467,6 @@ export function SearchModal({
     [getCurrentItem]
   )
 
-  // Render skeleton cards for loading state
   const renderSkeletonCards = () => {
     return Array.from({ length: 8 }).map((_, index) => (
       <div key={`skeleton-${index}`} className='w-80 flex-shrink-0'>
